refactor: cleanup the code
